Develop

+++ currency-conversion-service/src/main/java/com/ubaid/ms/currencyconversionservice/service/CurrencyConversionService.java
@@ -4,5 +4,5 @@ package com.ubaid.ms.currencyconversionservice.service;
 import com.ubaid.ms.ccdto.ConvertedCurrency;
 
 public interface CurrencyConversionService {
-    ConvertedCurrency convert(Integer currency, Integer conversionRate);
+    ConvertedCurrency convert(Double currency, Double conversionRate);
 }
