+++ rest-api/src/main/java/com/ubaid/ms/restapi/service/CurrencyConversionService.java
@@ -4,5 +4,5 @@ package com.ubaid.ms.restapi.service;
 import com.ubaid.ms.ccdto.ExchangeValueDTO;
 
 public interface CurrencyConversionService {
-    ExchangeValueDTO convertCurrency(String fromCurrency, String toCurrency, Integer quantity);
+    ExchangeValueDTO convertCurrency(String fromCurrency, String toCurrency, Double quantity);
 }
